requirements.txt
# StaffLens Dependencies
# Install with: pip install -r requirements.txt

# Discord (using py-cord for voice recording support)
py-cord[voice]>=2.6.0
PyNaCl>=1.5.0  # Required for voice support

# Environment & Configuration
python-dotenv>=1.0.0

# Text-to-Speech (for AI interviewer voice)
edge-tts>=6.1.0

# Transcription (using Deepgram REST API via aiohttp)
# No deepgram-sdk needed - we use direct HTTP requests

# HTTP Requests (for OpenRouter API & local endpoint)
aiohttp>=3.9.0

# Database
aiosqlite>=0.19.0

# FFmpeg required for audio playback - install separately:
# Windows: winget install ffmpeg  OR  choco install ffmpeg  OR  download from https://ffmpeg.org/download.html
# Mac: brew install ffmpeg
# Linux: apt install ffmpeg
